refactor(rust): extract db metadata operations (slice 6/9)
Move database, collection, and index metadata operations into db/metadata.rs:

- list_databases_impl
- list_collections_impl
- list_indexes_impl
- create_index_impl
- delete_index_impl

Keep crate-root re-exports so Tauri command wrappers and tests continue to call the same symbols. Behavior unchanged.

cargo check: passed
cargo test: 60 passed
